Long patch notes
I'm aiming for the patch notes for this single update to be larger than the patch notes of every other update combined. And with the amount of new things implemented in the VRAINS update, it's actually quite likely. I now have an interested mobile tester, so I may make note in this version or create a new mobile edition.